+++ deno/gateway/common.ts
@@ -1,5 +1,5 @@
 /**
- * https://discord.com/developers/docs/topics/gateway#connecting-gateway-url-params
+ * https://discord.com/developers/docs/topics/gateway#connecting-gateway-url-query-string-params
  */
 export interface GatewayURLQuery {
 	v: string;