+++ deno/payloads/v10/mod.ts
@@ -1,6 +1,7 @@
 export * from '../common.ts';
 export * from './application.ts';
 export * from './auditLog.ts';
+export * from './autoModeration.ts';
 export * from './channel.ts';
 export * from './emoji.ts';
 export * from './gateway.ts';