+++ rest/common.ts
@@ -179,6 +179,8 @@ export enum RESTJSONErrorCodes {
 	InvalidFormBodyOrContentType,
 	InviteAcceptedToGuildWithoutTheBotBeingIn,
 
+	InvalidActivityAction = 50039,
+
 	InvalidAPIVersion = 50041,
 
 	FileUploadedExceedsMaximumSize = 50045,
@@ -191,7 +193,8 @@ export enum RESTJSONErrorCodes {
 
 	PaymentSourceRequiredToRedeemGift = 50070,
 
-	CannotDeleteChannelRequiredForCommunityGuilds = 50074,
+	CannotModifyASystemWebhook = 50073,
+	CannotDeleteChannelRequiredForCommunityGuilds,
 
 	CannotEditStickersWithinMessage = 50080,
 	InvalidStickerSent,