+++ rest/v9/voice.ts
@@ -3,4 +3,9 @@ import type { APIVoiceRegion } from '../../payloads/v9/index';
 /**
  * https://discord.com/developers/docs/resources/voice#list-voice-regions
  */
-export type GetAPIVoiceRegionsResult = APIVoiceRegion[];
+export type RESTGetAPIVoiceRegionsResult = APIVoiceRegion[];
+
+/**
+ * @deprecated This was exported with the wrong name, use `RESTGetAPIVoiceRegionsResult` instead
+ */
+export type GetAPIVoiceRegionsResult = RESTGetAPIVoiceRegionsResult;